symbols: working on scaling and transforming correctly
