+++ icon/symbols/DoneSymbol/src/DoneSymbol.tsx
@@ -1,9 +1,8 @@
 import React, { ReactElement } from 'react';
-import { SymbolIcon } from '../../src/IconVariants';
-import { SpecializedSymbolProps } from '../types';
+import { SymbolIcon } from '../../../src/IconVariants';
+import { SpecializedSymbolProps } from '../../types';
 
-const path = <circle cx="6.00024" cy="6" r="6" fill="#11BDAF" />;
-
+const path = <circle cx="3.00024" cy="3" r="3" fill="#11BDAF" />;
 const DoneSymbol = ({ size = 'l', label }: SpecializedSymbolProps): ReactElement => {
   return <SymbolIcon label={label || 'done-symbol'} size={size} path={path} />;
 };
